+++ src/main/java/org/apostolis/comments/domain/Comment.java
@@ -5,13 +5,13 @@ import org.apostolis.AppConfig;
 import java.time.LocalDateTime;
 
 public record Comment(
-        int user,
-        int post,
+        long user,
+        long post,
         String text,
         LocalDateTime createdAt
 ){
 
-    public Comment(int user, int post, String text) {
+    public Comment(long user, long post, String text) {
         this(user,post,text,LocalDateTime.now(AppConfig.clock));
     }
 }
