+++ src/main/java/org/apostolis/comments/domain/CreateCommentRequest.java
@@ -1,4 +1,4 @@
 package org.apostolis.comments.domain;
 
-public record CreateCommentRequest(int post, String text) {
+public record CreateCommentRequest(long post, String text) {
 }
